docs(example-playwright): use fetch client
