+++ examples/with-playwright/package.json
@@ -19,7 +19,8 @@
     "clsx": "^2.1.1",
     "next": "^15.1.7",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "zimic": "latest"
   },
   "devDependencies": {
     "@playwright/test": "^1.50.1",
@@ -32,7 +33,6 @@
     "postcss": "^8.5.2",
     "tailwindcss": "^4.0.6",
     "tsx": "^4.19.2",
-    "typescript": "^5.7.3",
-    "zimic": "latest"
+    "typescript": "^5.7.3"
   }
 }
